+++ railties/lib/rails/generators/rails/app/app_generator.rb
@@ -206,7 +206,7 @@ module Rails
     end
 
     def database_yml
-      template "config/databases/#{options[:database]}.yml", "config/database.yml"
+      template database.template, "config/database.yml"
     end
 
     def db